backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.24.0
gunicorn>=21.2.0
python-dotenv>=1.0
psycopg2-binary>=2.9.0
passlib>=1.7.0
python-jose[cryptography]>=3.3.0
